requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic==2.5.0
streamlit>=1.28.0
requests>=2.31.0
python-multipart>=0.0.6
openai>=1.0.0
langchain==0.0.350
langchain-openai==0.0.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
duckduckgo-search==4.1.1
Pillow==10.1.0
streamlit-option-menu==0.3.6
pdfplumber==0.10.3
supabase==2.0.2 